refactor(sdk): extract messaging oneshot primitives to _client/_messaging_oneshot.py
PR 7/N of the SDK refactor split (see #702 WebSocket, #706 Guardian,
#711 Discovery, #713 RFQ, #716 AI gateway, #718 Sessions for the
mixin template). Moves the nine A2A oneshot/fire-and-forget
primitives — the hot domain of the SDK — out of cullis_sdk/client.py
into a dedicated mixin under cullis_sdk/_client/.

What moves:

- cullis_sdk/_client/_messaging_oneshot.py — new module exposing
  _MessagingOneshotMixin with:
    send_oneshot, reply_oneshot, receive_oneshot, decrypt_oneshot
      (# ── ADR-008 Phase 1 — sessionless one-shot ──)
    send_to_inbox, send_oneshot_and_wait
      (# ── ADR-020 Phase 4 — user inbox send ──)
    forward_oneshot, poll_oneshot_inbox, ack_oneshot
      (# ── Broker one-shot forwarding ──)
  Plus the scoped crypto imports: sign_message, sign_oneshot_envelope,
  verify_oneshot_envelope_signature, ONESHOT_ENVELOPE_PROTO_VERSION,
  encrypt_for_agent, decrypt_from_agent. Lazy imports inside method
  bodies (verify_inner_signature, json) preserved verbatim.

- cullis_sdk/client.py:
  * CullisClient now inherits
    (_AiGatewayMixin, _DiscoveryMixin, _GuardianMixin,
     _MessagingOneshotMixin, _RfqMixin, _SessionsMixin,
     _WebSocketMixin)
    in alphabetical order (A < D < G < Messag < R < S < W).
  * Nine methods, their three section headers, and the now-unused
    sign_oneshot_envelope, verify_oneshot_envelope_signature and
    ONESHOT_ENVELOPE_PROTO_VERSION imports are removed from client.py.
  * sign_message and verify_signature stay imported in client.py
    because the legacy session-based send/receive paths (which stay in
    client.py for now) still use them.
  * The _unwrap_egress_message staticmethod helper stays in client.py
    because it is used by the legacy poll() method, not by any oneshot
    method.

- cullis_sdk/__init__.py — unchanged. None of these methods were
  re-exported as top-level symbols.

- tests/test_sdk_guardian_phase3_send.py — updates the monkeypatch
  fixture paths from cullis_sdk.client.sign_* to the new canonical
  cullis_sdk._client._messaging_oneshot.sign_* location, since
  send_oneshot now resolves those names against the mixin module, not
  the client module. Pure plumbing follow-on, no test logic change.

Preserved behavior, explicitly checked:

- Crypto order on send: guardian-inspect(out) → encrypt → sign-outer.
  NEVER inverted — guardian-inspect-and-substitute happens BEFORE
  signing so the inner signature binds the redacted form.
- Crypto order on deliver: verify-outer → decrypt → verify-inner →
  guardian-inspect(in). The PubkeyFetchError / GuardianBlocked /
  ValueError exception paths are byte-identical.
- "oneshot:<corr>" session-id domain separation kept verbatim in both
  sign and decrypt paths.
- v=ONESHOT_ENVELOPE_PROTO_VERSION protocol version gate hard-rejects
  v1 envelopes unchanged.
- Cross-org trust-anchor delegation (issue #459) preserved: intra-org
  uses local _ca_chain_path, cross-org passes None and relies on
  Court-side chain verification at federation publish time.
- send_oneshot_and_wait polling loop and reply matching by
  correlation_id preserved verbatim. The "no polling autonomous on
  inbox" behavioural memory rule applies to agent runtime usage, not
  to this SDK method's contract.

No logic change — pure code movement (modulo the test fixture path
update, which is a plumbing follow-on of the move itself).

No @deprecated annotations in this PR — the eventual deprecation of
the session-based legacy path will land in a dedicated PR after the
client.py split completes.
